pi-code/requirements.txt
requests
configparser
cryptography
pyopenssl
PyJWT
paho-mqtt
Dumper
Flask
uwsgi
pytz
tzlocal
